added dropdowncomponent to NN page, changed buttons to download links  and put NN in screens folder
